Add moshi library to prepare parsing json config files
This is in preparation to allow users to load their own config file to
the LogAnalyzer without needing to rebuild it.

Reasearch found Moshi and Okio libraries had the right amount of
performance while also allowing the json config file to be both strict
on checking required fields with type checks, while being flexible
enough to extra fields, while also giving good error messages.

The animal example is just for future reference, it will be deleted in
the next commit